+++ tap_dayforce/__init__.py
@@ -1,85 +1,95 @@
-import logging
+import json
 import os
 
 import rollbar
 import singer
-from rollbar.logger import RollbarHandler
 
-from .streams import AVAILABLE_STREAMS, ReportStream
+from .streams import (EmployeePunchesStream, EmployeeRawPunchesStream,
+                      EmployeesStream, PaySummaryReportStream)
+from .utils import load_schema, parse_args
 
-ROLLBAR_ACCESS_TOKEN = os.environ["ROLLBAR_ACCESS_TOKEN"]
-ROLLBAR_ENVIRONMENT = os.environ["ROLLBAR_ENVIRONMENT"]
+AVAILABLE_STREAMS = {
+    EmployeePunchesStream,
+    EmployeeRawPunchesStream,
+    EmployeesStream,
+    PaySummaryReportStream
+}
 
 LOGGER = singer.get_logger()
 
-rollbar.init(ROLLBAR_ACCESS_TOKEN, ROLLBAR_ENVIRONMENT)
-rollbar_handler = RollbarHandler()
-# Only send level WARNING and above to Rollbar.
-rollbar_handler.setLevel(logging.WARNING)
-LOGGER.addHandler(rollbar_handler)
+try:
+    ROLLBAR_ACCESS_TOKEN = os.environ["ROLLBAR_ACCESS_TOKEN"]
+    ROLLBAR_ENVIRONMENT = os.environ["ROLLBAR_ENVIRONMENT"]
+except KeyError:
+    LOGGER.info("Required env vars for Rollbar logging not found. Rollbar logging disabled..")
+    log_to_rollbar = False
+else:
+    rollbar.init(ROLLBAR_ACCESS_TOKEN, ROLLBAR_ENVIRONMENT)
+    log_to_rollbar = True
 
 
-def discover(config, state={}):
+def discover(args, select_all=False):
     LOGGER.info('Starting discovery..')
-    data = {}
-    data['streams'] = []
-    for available_stream in AVAILABLE_STREAMS:
-        if available_stream == ReportStream:
-            if config.get("streams", {}).get("reports") is None:
-                LOGGER.info('No Reports to replicate. Moving on..')
-                continue
-            else:
-                for report in config.get("streams").get("reports").keys():
-                    data['streams'].append(available_stream(config=config, state=state, xrefcode=report))
-        else:
-            data['streams'].append(available_stream(config=config, state=state))
-    catalog = singer.catalog.Catalog.from_dict(data=data)
-    singer.catalog.write_catalog(catalog)
+
+    catalog = {"streams": []}
+    for stream in AVAILABLE_STREAMS:
+        schema = load_schema(stream.tap_stream_id)
+        catalog_entry = {
+            "stream": stream.tap_stream_id,
+            "tap_stream_id": stream.tap_stream_id,
+            "schema": schema,
+            "metadata": singer.metadata.get_standard_metadata(schema=schema,
+                                                              key_properties=stream.key_properties,
+                                                              valid_replication_keys=stream.bookmark_properties,
+                                                              replication_method=stream.replication_method)
+        }
+
+        if select_all is True:
+            catalog_entry["metadata"][0]["metadata"]["selected"] = True
+        catalog["streams"].append(catalog_entry)
+
+    print(json.dumps(catalog, indent=2))
     LOGGER.info('Finished discovery..')
 
 
-def sync(config, catalog, state):
+def sync(args):
     LOGGER.info('Starting sync..')
-    selected_streams = {catalog_entry.stream for catalog_entry in catalog.get_selected_streams(state)}
 
-    streams_to_sync = set()
+    selected_streams = {catalog_entry.stream for catalog_entry in args.catalog.get_selected_streams(args.state)}
+    LOGGER.info(f"Selected Streams: {selected_streams}")
+
     for available_stream in AVAILABLE_STREAMS:
-        if available_stream == ReportStream:
-            if config.get("streams", {}).get("reports") is None:
-                LOGGER.info('No Reports to replicate. Moving on..')
-                continue
-            else:
-                for report in config.get("streams").get("reports").keys():
-                    stream = available_stream(config=config, state=state, xrefcode=report)
-                    if stream.stream in selected_streams:
-                        streams_to_sync.add(stream)
-
-        elif available_stream.stream in selected_streams:
-            streams_to_sync.add(available_stream(config=config, state=state))
-
-    for stream in streams_to_sync:
-        singer.bookmarks.set_currently_syncing(state=stream.state, tap_stream_id=stream.tap_stream_id)
-        stream.write_state_message()
-        stream.write_schema_message()
-        stream.sync()
-        singer.bookmarks.set_currently_syncing(state=stream.state, tap_stream_id=None)
-        stream.write_state_message()
+        stream = available_stream.from_args(args)
+        if stream.tap_stream_id in selected_streams:
+            LOGGER.info(f"Starting sync for Stream {stream.tap_stream_id}..")
+            singer.bookmarks.set_currently_syncing(state=stream.state, tap_stream_id=stream.tap_stream_id)
+            singer.write_state(stream.state)
+            singer.write_schema(stream_name=stream.tap_stream_id,
+                                schema=stream.get_schema(tap_stream_id=stream.tap_stream_id, catalog=stream.catalog),
+                                key_properties=stream.key_properties)
+            stream.sync()
+            singer.bookmarks.set_currently_syncing(state=stream.state, tap_stream_id=None)
+            singer.write_state(stream.state)
 
 
 def main():
-    args = singer.utils.parse_args(required_config_keys=["username", "password", "client_name", "email"])
+    args = parse_args(required_config_keys={"username", "password", "client_namespace", "start_date"})
     if args.discover:
         try:
-            discover(config=args.config)
+            discover(args, select_all=args.select_all)
         except:
             LOGGER.exception('Caught exception during Discovery..')
-            rollbar.report_exc_info()
+            if log_to_rollbar is True:
+                rollbar.report_exc_info()
+    elif not args.catalog:
+        raise RuntimeError('Catalog file must be supplied during Sync.')
     else:
         try:
-            sync(config=args.config, catalog=args.catalog, state=args.state)
+            sync(args)
         except:
             LOGGER.exception('Caught exception during Sync..')
-            rollbar.report_exc_info()
+            if log_to_rollbar is True:
+                rollbar.report_exc_info()
 
 
 if __name__ == "__main__":